+++ src/kernels/backend_webgl.ts
@@ -101,6 +101,8 @@ export interface WebGLTimingInfo extends TimingInfo {
   downloadWaitMs: number;
 }
 
+// Combines a dataId, a shape, and a dtype without a Tensor object so that
+// programs can be executed without a full Tensor object.
 export interface TensorHandle {
   dataId: DataId;
   shape: number[];
